coalesce each DL algorithm implementation into one file
