+++ src/bl12/mod.rs
@@ -1,21 +1,30 @@
 #![allow(non_snake_case)]
 
-pub mod generator;
+//! [BL12] discrete logarithm solver.
+//!
+//! Implements the algorithm from "Computing small discrete logarithms faster"
+//! by Daniel J. Bernstein and Tanja Lange (2012).
+//! https://cr.yp.to/dlog/cuberoot-20120919.pdf
+
 #[cfg(feature = "bl12_presets")]
 pub mod presets;
-pub mod solver;
 
 #[cfg(feature = "bl12_presets")]
 use crate::bl12::presets::Presets;
+use crate::utils;
 
 use anyhow::{Context, Result};
+use curve25519_dalek::constants::RISTRETTO_BASEPOINT_POINT;
 use curve25519_dalek::ristretto::{CompressedRistretto, RistrettoPoint};
 use curve25519_dalek::scalar::Scalar;
+use curve25519_dalek::traits::Identity;
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
 #[cfg(feature = "serde")]
 use serde_with::serde_as;
 use std::collections::HashMap;
+use std::ops::{Add, AddAssign, Mul, Sub};
+use web_time::{Duration, Instant};
 
 /// [BL12] discrete logarithm solver.
 ///
@@ -93,12 +102,144 @@ impl Bl12 {
 
         Ok(bl12)
     }
+
+    pub fn solve_dlp(&self, pk: &RistrettoPoint, max_time: Option<u64>) -> Result<Option<u64>> {
+        if pk.eq(&RistrettoPoint::identity()) {
+            return Ok(Some(0));
+        }
+
+        let start_time = max_time.map(|_| Instant::now());
+
+        loop {
+            // wdist = r + slog_1 + slog_2 ...
+            // For small secret sizes, use 0 or a small value
+            let wdist_bits = self.parameters.secret_size.saturating_sub(8).max(1);
+            let mut wdist = utils::generate_random_scalar(wdist_bits)
+                .context("failed to generate `wdist` scalar")?;
+            // w = sk * G + r * G + slog_1 * G + slog_2 * G ... = sk * G + wdist * G
+            let mut w = pk.add(RISTRETTO_BASEPOINT_POINT.mul(wdist));
+
+            for _ in 0..self.parameters.i * self.parameters.W {
+                let w_compressed = w.compress();
+
+                if is_distinguished(&w_compressed, &self.parameters) {
+                    if let Some(value) = self.table.table.get(&w_compressed) {
+                        // value * G = sk * G + wdist * G => sk = value - wdist
+                        let sk = value.sub(wdist);
+
+                        assert!(RISTRETTO_BASEPOINT_POINT.mul(sk).eq(pk));
+
+                        return Ok(Some(utils::scalar_to_u64(&sk)));
+                    }
+
+                    break;
+                }
+
+                if let Some(max_time) = max_time {
+                    if start_time.unwrap().elapsed() >= Duration::from_millis(max_time) {
+                        return Ok(None);
+                    }
+                }
+
+                let h = hash(&w_compressed, &self.parameters) as usize;
+
+                wdist.add_assign(&self.table.slog[h]);
+                w.add_assign(&self.table.s[h]);
+            }
+        }
+    }
 }
 
-pub(crate) fn is_distinguished(
-    compressed_point: &CompressedRistretto,
-    parameters: &Parameters,
-) -> bool {
+impl Table {
+    pub fn generate(parameters: &Parameters) -> Result<Table> {
+        if parameters.secret_size < 1 || parameters.secret_size > 64 {
+            return Err(anyhow::anyhow!("secret size must be between 1 and 64"));
+        }
+
+        let (slog, s) = Self::s_values_init(parameters)?;
+
+        let mut table = HashMap::new();
+
+        // Track attempts to detect if we're stuck finding the same distinguished points
+        let mut attempts = 0u64;
+        let max_attempts = parameters.N as u64 * 1000;
+
+        while table.len() < parameters.N as usize {
+            attempts += 1;
+            if attempts > max_attempts {
+                // Give up if we can't find enough unique distinguished points.
+                // This can happen for very small search spaces.
+                break;
+            }
+
+            let mut wlog = utils::generate_random_scalar(parameters.secret_size)
+                .context("failed to generate `wlog` scalar")?;
+            let mut w = RISTRETTO_BASEPOINT_POINT.mul(wlog);
+
+            for _ in 0..parameters.i * parameters.W {
+                let w_compressed = w.compress();
+
+                if is_distinguished(&w_compressed, parameters) {
+                    table.insert(w_compressed, wlog);
+
+                    break;
+                }
+
+                let h = hash(&w_compressed, parameters) as usize;
+
+                wlog.add_assign(slog[h]);
+                w.add_assign(s[h]);
+            }
+        }
+
+        Ok(Table { s, slog, table })
+    }
+
+    fn s_values_init(parameters: &Parameters) -> Result<(Vec<Scalar>, Vec<RistrettoPoint>)> {
+        // Calculate slog_size: step sizes should allow walks to cover the search space
+        // in roughly W steps (the expected distance between distinguished points).
+        //
+        // For small secret_size (< 8), use secret_size directly as slog_size so that
+        // steps can span the entire search space. This ensures the random walk can
+        // effectively explore small spaces.
+        //
+        // For larger secret_size, use the ratio-based calculation which gives
+        // step sizes proportional to search_space / (4 * W).
+        let slog_size = if parameters.secret_size < 8 {
+            parameters.secret_size.max(1)
+        } else {
+            let search_space = 1u64 << parameters.secret_size.min(62);
+            let ratio = search_space
+                .saturating_div(4)
+                .saturating_div(parameters.W.max(1));
+            if ratio > 0 {
+                ratio.ilog2().max(1) as u8
+            } else {
+                1
+            }
+        };
+
+        let mut scalars = Vec::with_capacity(parameters.R as usize);
+        let mut points = Vec::with_capacity(parameters.R as usize);
+
+        for _ in 0..parameters.R {
+            // Generate step size in range [1, 2^slog_size] to ensure non-zero steps
+            // and good variety. We generate a random value and add 1.
+            let random_part = utils::generate_random_scalar(slog_size)
+                .context("failed to generate `slog` scalar")?;
+            let slog = random_part + Scalar::ONE;
+
+            let s = RISTRETTO_BASEPOINT_POINT.mul(slog);
+
+            scalars.push(slog);
+            points.push(s);
+        }
+
+        Ok((scalars, points))
+    }
+}
+
+fn is_distinguished(compressed_point: &CompressedRistretto, parameters: &Parameters) -> bool {
     let point_bytes = get_last_point_bytes(compressed_point);
 
     (point_bytes & (parameters.W - 1)) == 0
@@ -109,7 +250,7 @@ pub(crate) fn is_distinguished(
 ///
 /// Note: it does not perform hashing. However, in the original reference implementation authors
 /// (Daniel J. Bernstein and Tanja Lange) use exactly the same name.
-pub(crate) fn hash(compressed_point: &CompressedRistretto, parameters: &Parameters) -> u64 {
+fn hash(compressed_point: &CompressedRistretto, parameters: &Parameters) -> u64 {
     let point_bytes = get_last_point_bytes(compressed_point);
 
     point_bytes & (parameters.R - 1)
